readme
Description of research and files.

This repository contains documents related to labeling, filtering, and machine learning classification of tweets describing Covid-19-related behaviors from five states: Michigan, New York, Idaho, Alabama, and Florida. Tweets with at least one keyword/phrase that describes a safer behavior (like staying home) are referred to as “increased”. Tweets describing less safe behaviors (like going to the gym) are referred to as “decreased”.

The repository is organized into folders based on steps in the process, or the type of data contained within. Please also refer to the “fall_ind_study_write_up” document, as this makes reference to different steps detailed in the research.