Write up of independent study
The work aims to create two machine learning classifiers designed to classify tweets related to the COVID-19 pandemic. This project aims to detect tweets that represent self-reported behaviors that have increased that represent safer behaviors (referred to in this paper as “increased”) and behaviors that may have decreased that represent less safe behaviors (referred to as “decreased”). 

The goal of this research is to develop a technique to identify tweets that describe a behavior, and later, plotting the change in relative frequency of tweets that are related to the pandemic in a time-series graph.