Activity Tweet Annotation Guide
An activity tweet is one that describes performing an action, refraining from performing an action, or encouraging or discouraging people from performing the action. It also can be related to someone being at a location (e.g. “office”). It should directly relate to the activity keyword(s) listed in the column text_keywords in each spreadsheet. The tweet can involve the user, someone the user is with, or be an observation about something that is/isn’t happening in their immediate surroundings.

This is a guide used to noted "increased" and "decreased" activities related to the COVID-19 pandemic.